Read timestamp as f64 directly to preserve millisecond precision
Previously timestamps were read as i64 then cast to f64, which would
truncate any fractional seconds (millisecond precision) stored in the
parquet files. Now reads directly as f64 to preserve full precision.

+++ applications/datamanager/src/storage.rs
@@ -565,7 +565,7 @@ fn execute_portfolio_query_with_action(
         .query_map([], |row| {
             Ok(Portfolio {
                 ticker: row.get::<_, String>(0)?,
-                timestamp: row.get::<_, i64>(1)? as f64,
+                timestamp: row.get::<_, f64>(1)?,
                 side: row.get::<_, String>(2)?,
                 dollar_amount: row.get::<_, f64>(3)?,
                 action: row.get::<_, String>(4)?,
@@ -592,7 +592,7 @@ fn execute_portfolio_query_without_action(
         .query_map([], |row| {
             Ok(Portfolio {
                 ticker: row.get::<_, String>(0)?,
-                timestamp: row.get::<_, i64>(1)? as f64,
+                timestamp: row.get::<_, f64>(1)?,
                 side: row.get::<_, String>(2)?,
                 dollar_amount: row.get::<_, f64>(3)?,
                 action: "UNSPECIFIED".to_string(),
